# farol/docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Farol API",
    "description": "Gestão, priorização e despacho de ordens de serviço elétricas. Dados e pontuações demonstrativos não são oficiais.",
    "version": "v1"
  },
  "paths": {
...
    "/api/municipios": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "todos",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MunicipioDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MunicipioDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MunicipioDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/municipios/{id}/subestacoes": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubestacaoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubestacaoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubestacaoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/subestacoes/{id}/conjuntos": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConjuntoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConjuntoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConjuntoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/localidades": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalidadeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalidadeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalidadeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/transformadores": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numero",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanhoPagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransformadorDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformadorDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformadorDtoPaginaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/transformadores/interpretar": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "summary": "Interpreta o número: 3 dígitos de localidade + número local; sugere circuito e conjunto do cadastro.",
        "parameters": [
          {
            "name": "numero",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretacaoTransformadorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretacaoTransformadorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterpretacaoTransformadorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/unidades-consumidoras": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "summary": "Busca de UC — consulta auditada, nome mascarado.",
        "parameters": [
          {
            "name": "uc",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeConsumidoraDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeConsumidoraDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeConsumidoraDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/geocodificacao/cep/{cep}": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "cep",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CepDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CepDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CepDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/geocodificacao/coordenadas/interpretar": {
      "post": {
        "tags": [
          "Cadastros"
        ],
        "summary": "Aceita par decimal, link de mapa ou graus/minutos/segundos.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretarCoordenadasRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretarCoordenadasRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretarCoordenadasRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoordenadasInterpretadasDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordenadasInterpretadasDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordenadasInterpretadasDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/tipos-ocorrencia": {
      "get": {
        "tags": [
          "Cadastros"
        ],
        "parameters": [
          {
            "name": "todos",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TipoOcorrenciaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TipoOcorrenciaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TipoOcorrenciaDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/configuracoes/pontuacao": {
      "get": {
        "tags": [
          "Pontuacao"
        ],
        "summary": "Versão vigente, rascunho e versão aguardando aprovação do escopo (global ou cidade).",
        "parameters": [
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracaoPontuacaoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracaoPontuacaoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracaoPontuacaoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/rascunho": {
      "post": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pontuacao"
        ],
        "summary": "Grava o rascunho. Gravação concorrente com token antigo recebe 409.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarRascunhoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarRascunhoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarRascunhoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/rascunho/{versaoId}": {
      "delete": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "versaoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/simular": {
      "post": {
        "tags": [
          "Pontuacao"
        ],
        "summary": "Simulador: mesmo motor da produção, com o rascunho ou uma versão escolhida.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimularRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SimularRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SimularRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificacaoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificacaoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificacaoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/impacto": {
      "get": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "versaoId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactoVersaoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactoVersaoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactoVersaoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/publicar": {
      "post": {
        "tags": [
          "Pontuacao"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicarRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicarRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PublicarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/versoes/{id}/aprovar": {
      "post": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/versoes/{id}/rejeitar": {
      "post": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "motivo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/versoes": {
      "get": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersaoListaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersaoListaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersaoListaDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/versoes/{id}": {
      "get": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/versoes/comparar": {
      "get": {
        "tags": [
          "Pontuacao"
        ],
        "parameters": [
          {
            "name": "a",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "b",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComparacaoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparacaoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparacaoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracoes/pontuacao/versoes/{id}/restaurar": {
      "post": {
        "tags": [
          "Pontuacao"
        ],
        "summary": "Restaurar cria um novo rascunho com o conteúdo da versão; o histórico fica intacto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "municipioId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersaoDetalheDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConjuntoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "subestacaoId": {
            "type": "integer",
            "format": "int32"
          },
          "numero": {
            "type": "string",
            "nullable": true
          },
          "demonstrativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "MunicipioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "uf": {
            "type": "string",
            "nullable": true
          },
          "codigoIbge": {
            "type": "string",
            "nullable": true
          },
          "prefixo": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "raioKm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cepUnico": {
            "type": "string",
            "nullable": true
          },
          "ativo": {
            "type": "boolean"
          },
          "demonstrativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Token obtido em POST /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}